requirements.txt
pytest
selenium
pyyaml
